pyproject.toml
[project]
name = "visualvoice"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "setuptools>=68.0.0",
    "opencv-python<=4.8.0.74",
    "librosa==0.7.0",
    "imageio<=2.31.1",
    "scipy<=1.10.1",
    "Pillow<=10.0.0",
    "facenet-pytorch<=2.5.3",
    "torch==1.7.1+cu110",
    "matplotlib<=3.7.2",
    "face-alignment==1.2.0",
    "numba==0.48",
    "mmcv<=2.0.1",
    "h5py<=3.9.0",
    "av==8.0.3",
    "resampy>=0.2.0,<0.4.2",
    "llvmlite<=0.33.0",
    "numpy<=1.23.0",
    "joblib>=1.3.1",
    "soundfile>=0.12.1",
    "ffmpeg>=1.4",
    "ffmpeg-python>=0.2.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.4.1",
    "flake8>=5.0.4",
    "pytest>=7.4.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[[tool.rye.sources]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu110"
type = "index"

[tool.rye.scripts]
split = { call = "collection.collect_samples:cli" }
filter = { call = "collection.filter_samples:cli" }
